subtests.docker_cli.kill: Add support for "os.kill" and attached process
This pull request adds the possibility to use os.kill instead of
"docker kill" as a test command.

Additionally it adds the possibility to run the main docker command as
docker attach process.

Booth features are utilized in 4 new tests, which test the sig-proxy
feature of the main process (run or attach).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>