dockertest: Add support for (Sub)Subtest.manager
(Sub)Subtest.manager is a special class, which helps managing containers
and other objects created in test. Currently it supports creating a named
container, storing containers and images and cleanup of created/registered
containers (there is class for images too, but it is not implemented)

Additionally this patch modifies existing tests to work fine with this
modification.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>